+++ node/src/main/java/CordaCaplet.java
@@ -93,6 +93,7 @@ public class CordaCaplet extends Capsule {
             (new File(baseDir, "cordapps")).mkdir();
             // Add additional directories of JARs to the classpath (at the end). e.g. for JDBC drivers
             augmentClasspath((List<Path>) cp, new File(baseDir, "drivers"));
+            augmentClasspath((List<Path>) cp, new File(baseDir, "cordapps"));
             try {
                 List<String> jarDirs = nodeConfig.getStringList("jarDirs");
                 log(LOG_VERBOSE, "Configured JAR directories = " + jarDirs);